ex 4.21*

+++ osa4/blogilista/controllers/blogs.js
@@ -2,6 +2,7 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
+require('express-async-errors')
 
 
 blogsRouter.get('/', async (request, response) => {
@@ -10,39 +11,40 @@ blogsRouter.get('/', async (request, response) => {
 
 })
 
-blogsRouter.get('/:id', async (request, response, next) => {
-  try {
-    const blog = await Blog.findById(request.params.id)
-    if (blog) {
-      response.json(blog)
-    } else {
-      response.status(404).end()
-    }
-  } catch (exception) {
-    next(exception)
+blogsRouter.get('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (blog) {
+    response.json(blog)
+  } else {
+    response.status(404).end()
   }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
-  try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
-  } catch (exception) {
-    next(exception)
+blogsRouter.delete('/:id', async (request, response) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
   }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({ error: 'invalid credentials' })
+  }
+
+  await Blog.findByIdAndRemove(request.params.id)
+  response.status(204).end()
 })
 
-blogsRouter.put('/:id', async (request, response, next) => {
-  try {
-    await Blog.findByIdAndUpdate(request.params.id, request.body)
-    response.status(200).end()
-  } catch (exception) {
-    next(exception)
-  }
+
+blogsRouter.put('/:id', async (request, response) => {
+  await Blog.findByIdAndUpdate(request.params.id, request.body)
+  response.status(200).end()
 })
 
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', async (request, response) => {
 
   if (request.body.title === undefined && request.body.url === undefined) {
     response.status(400).end()
@@ -64,15 +66,11 @@ blogsRouter.post('/', async (request, response, next) => {
     user: user._id
   })
 
-  try {
-    const savedBlog = await blog.save()
-    user.blogs = user.blogs.concat(savedBlog._id)
-    await user.save()
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
 
-    response.status(201).json(savedBlog)
-  } catch (exception) {
-    next(exception)
-  }
+  response.status(201).json(savedBlog)
 })
 
 module.exports = blogsRouter
